requirements.txt
dnspython
tabulate